+++ deepmatch-testing/src/main/kotlin/com/aouledissa/deepmatch/processor/FakeDeeplinkProcessor.kt
@@ -1,9 +1,20 @@
 package com.aouledissa.deepmatch.processor
 
-import android.app.Activity
 import android.net.Uri
+import com.aouledissa.deepmatch.api.DeeplinkParams
 
 class FakeDeeplinkProcessor : DeeplinkProcessor {
-    override fun match(deeplink: Uri, activity: Activity) {
+    private var resultProvider: (Uri) -> DeeplinkParams? = { null }
+    private var lastDeeplink: Uri? = null
+
+    fun configure(resultProvider: (Uri) -> DeeplinkParams?) {
+        this.resultProvider = resultProvider
     }
+
+    override fun match(deeplink: Uri): DeeplinkParams? {
+        lastDeeplink = deeplink
+        return resultProvider(deeplink)
+    }
+
+    fun isInvoked() = lastDeeplink != null
 }